python-service/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
prometheus-fastapi-instrumentator
psycopg2
opentelemetry-instrumentation-psycopg2